backend/requirements.txt
feedparser
newspaper3k
pysolr
